+++ contentstack/entry.py
@@ -123,6 +123,23 @@ class Entry(EntryQueryable):
         self.entry_param['include_fallback'] = 'true'
         return self
 
+    def include_branch(self):
+        """Retrieve the published pranch in the entry response
+        :return: Entry, so we can chain the call
+        ----------------------------
+        Example::
+
+            >>> import contentstack
+            >>> stack = contentstack.Stack('api_key', 'delivery_token', 'environment')
+            >>> content_type = stack.content_type('content_type_uid')
+            >>> entry = content_type.entry(uid='entry_uid')
+            >>> entry = entry.include_branch()
+            >>> result = entry.fetch()
+        ----------------------------
+        """
+        self.entry_param['include_branch'] = 'true'
+        return self
+
     def include_embedded_items(self):
         """include_embedded_items instance of Entry
         include_embedded_objects (Entries and Assets) along with entry/entries details.